some details fixed

+++ ex01/ZombieHorde.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ZombieHorde.cpp                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kali <kali@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: anurtiag <anurtiag@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/08 04:40:57 by kali              #+#    #+#             */
-/*   Updated: 2024/06/08 05:18:42 by kali             ###   ########.fr       */
+/*   Updated: 2024/07/08 15:20:34 by anurtiag         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 Zombie *zombieHorde(int N, std::string name)
 {
     Zombie *zombie = new Zombie[N];
-    for(size_t i = 0; i < N; i++)
+    for(int i = 0; i < N; i++)
        zombie[i] = Zombie(name);
     return(zombie);
 }
